requirements-dev.txt
google-cloud
google-cloud-storage
pytest
pytest-watch